NN-2445 Tidy up dto and remove unneeded files
