--- src/main/java/net/syscon/elite/repository/jpa/repository/ScheduledAppointmentProjection.java
@@ -1,7 +1,0 @@
-package net.syscon.elite.repository.jpa.repository;
-
-public interface ScheduledAppointmentProjection {
-    String getOffenderNo();
-    String getFirstName();
-    String getLastName();
-}
